Refactored token auth
