+++ routes/public/user.go
@@ -10,7 +10,7 @@ import (
 func RegisterUser(e *gin.Engine) {
 	userGroup := e.Group("/user")
 
-	userGroup.GET("/register", controller.GetRegisterPage)
-	userGroup.GET("/login", controller.GetLoginPage)
-	userGroup.GET("/profile", middleware.VerifyAuthToken, controller.GetProfilePage)
+	userGroup.GET("/register", middleware.CreateAuthMiddleware(false), controller.GetRegisterPage)
+	userGroup.GET("/login", middleware.CreateAuthMiddleware(false), controller.GetLoginPage)
+	userGroup.GET("/profile", middleware.CreateAuthMiddleware(true), controller.GetProfilePage)
 }
